extract dataclasses for heat2030
